Experimental unit reconstruction

+++ src/CrossyToad/Effect/Time/Seconds.hs
@@ -1,3 +1,3 @@
 module CrossyToad.Effect.Time.Seconds where
 
-type Seconds = Float
+type Seconds = Int
